Reuse logic from `no-alphabetical-sort`
